refactor: created /src /include and /test dir for better structure & updated CMakeLists.txt
