+++ include/manager.h
